% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfunctions.R
\name{bft}
\alias{bft}
\title{Cumulant (log partition) function of cobin}
\usage{
bft(x)
}
\arguments{
\item{x}{input vector}
}
\value{
\eqn{B(x) = \log\{(\exp(x)-1)/x)\}}
}
\description{
\eqn{B(x) = \log\{(\exp(x)-1)/x)\}}
}
