% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfunctions.R
\name{bftprime}
\alias{bftprime}
\title{First derivative of cobin cumulant (log partition) function}
\usage{
bftprime(x)
}
\arguments{
\item{x}{}
}
\description{
B'(x) = 1/(1-exp(-x))-1/x.
When g is canonical link, this is same as g^(-1)
}
