% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfunctions.R
\name{bftprimeinv}
\alias{bftprimeinv}
\title{Inverse of first derivative of cobin cumulant (log partition) function}
\usage{
bftprimeinv(y, x0 = 0, tol = 1e-08, maxiter = 100)
}
\arguments{
\item{x0}{initial value}

\item{tol}{tolerance, stopping criterion for Newton-Raphson}

\item{maxiter}{max iteration of Newton-Raphson}
}
\description{
Calculates (B')^(-1)(y) using numerical inversion (Newton-Raphson)
This is the cobit link function g, the canonical link function of cobin
}
