% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfunctions.R
\name{bftprimeprime}
\alias{bftprimeprime}
\title{Second derivative of cobin cumulant (log partition) function}
\usage{
bftprimeprime(x)
}
\arguments{
\item{x}{}
}
\description{
B''(x) = 1/x^2 + 1/(2-2*cosh(x))
used Taylor series expansion for x near 0 for stability
When g is canonical link, this is same as (g^(-1))'
}
