% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varfunctions.R
\name{bftprimeprimeprime}
\alias{bftprimeprimeprime}
\title{Third derivative of cobin cumulant (log partition) function}
\usage{
bftprimeprimeprime(x)
}
\arguments{
\item{x}{}
}
\description{
B'''(x) = 1/(4*tanh(x/2)*sinh(x/2)^2) - 2/x^3
used Taylor series expansion for x near 0 for stability
}
