% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobinfamily.R
\name{cobinfamily}
\alias{cobinfamily}
\title{cobin family class}
\usage{
cobinfamily(lambda = stop("'lambda' must be specified"), link = "cobit")
}
\arguments{
\item{lambda}{The known value of lambda, must be integer}

\item{link}{The link function to be used. Options are "cobit" (canonical link for cobin regression), "logit", "probit", "cauchit", "cloglog"}
}
\value{
An object of class "family", a list of functions and expressions needed by glm() to fit a cobin generalized linear model.
}
\description{
Specifies the information required to fit a cobin generalized linear model with known lambda parameter, using glm().
}
