% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobindist.R
\name{dcobin}
\alias{dcobin}
\title{Density function of cobin (continuous binomial) distribution}
\usage{
dcobin(x, theta, lambda, log = FALSE)
}
\arguments{
\item{x}{num (length n), between 0 and 1, evaluation point}

\item{theta}{num (length 1 or n), canonical parameter}

\item{lambda}{integer, inverse of dispersion parameter}

\item{log}{logical, return log density?}
}
\description{
Density function of cobin (continuous binomial) distribution
}
\examples{

xgrid = seq(0, 1, length = 500)
plot(xgrid, dcobin(xgrid, 0, 1), type="l", ylim = c(0,3)) # uniform 
lines(xgrid, dcobin(xgrid, 0, 3))
plot(xgrid, dcobin(xgrid, 2, 3), type="l")
lines(xgrid, dcobin(xgrid, -2, 3))

}
