% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IH.R
\name{dIH}
\alias{dIH}
\title{Density of Irwin-Hall distribution}
\usage{
dIH(x, m, log = F)
}
\arguments{
\item{x}{num(length n), between 0 and m, evaluation point}

\item{m}{integer, parameter}

\item{log}{logical, return log density if TRUE}
}
\value{
density value at x
}
\description{
Irwin-Hall distribution with parameter m is defined as a sum of m uniform (0,1) distribution.
}
\details{

}
\examples{

m = 4
xgrid= seq(0, m, length = 500)
plot(xgrid, dIH(xgrid, m, log = FALSE))

}
