% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micobindist.R
\name{dmicobin}
\alias{dmicobin}
\title{Title}
\usage{
dmicobin(x, theta, psi, r = 2, log = FALSE, l_max = 70)
}
\arguments{
\item{x}{}

\item{theta}{}

\item{psi}{}

\item{r}{}

\item{log}{}

\item{l_max}{}
}
\description{
Title
}
