% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.cobin.R
\name{glm.cobin}
\alias{glm.cobin}
\title{Title}
\usage{
glm.cobin(
  formula,
  data,
  weights,
  subset,
  na.action,
  start = NULL,
  etastart,
  mustart,
  control = glm.control(...),
  method = "glm.fit",
  model = TRUE,
  x = FALSE,
  y = TRUE,
  contrasts = NULL,
  ...,
  lambda_list = 1:70,
  link = "cobit",
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\description{
Title
}
