% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobindist.R
\name{pcobin}
\alias{pcobin}
\title{Cumulative distribution function of cobin (continuous binomial) distribution}
\usage{
pcobin(q, theta, lambda)
}
\arguments{
\item{q}{num (length n), between 0 and 1, evaluation point}

\item{theta}{num 1, canonical parameter}

\item{lambda}{integer, inverse of dispersion parameter}
}
\description{
Cumulative distribution function of cobin (continuous binomial) distribution
}
