% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB.R
\name{qcb}
\alias{qcb}
\title{Quantile function of continuous Bernoulli (continuous binomial with lambda = 1)}
\usage{
qcb(p, theta)
}
\arguments{
\item{p}{length n vector, probabilities}

\item{theta}{scalar or length n vector, theta}
}
\value{
length n vector of quantiles
}
\description{
Continuous Bernoulli distribution with parameter theta has a density function
f(y; theta) = theta/(e^theta-1) * e^(theta * y)
or equivalently, with phi = e^theta/(1+e^theta),
f(y; phi) \propto phi^y * (1-phi)^(1-y)
}
\details{
Its quantile function is

log((exp(theta)-1)*p+1)/theta

below codes evalualtes in a numerically stable way
}
\examples{

pgrid = seq(0.001, 0.999, length = 1000)
plot(pgrid, qcb(pgrid, 0)) # quantile function of uniform
plot(pgrid, qcb(pgrid, 3)) #

}
