% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobindist.R
\name{rcobin}
\alias{rcobin}
\title{Random variate generation for cobin (continuous binomial) distribution}
\usage{
rcobin(n, theta, lambda)
}
\arguments{
\item{n}{integer, number of samples}

\item{theta}{scalar or length n vector, theta}

\item{lambda}{scalar or length n vector, lambda, length should be same as theta}
}
\description{
y ~ cobin(theta, 1/lambda)
which is distributionally equivalent to
x_1 + ... + x_lambda, x_l ~ contiBernoulli(theta)
}
\details{
Continuous Bernoulli distribution with parameter theta has a density function
f(y; theta) = theta/(e^theta-1) * e^(theta * y)
or equivalently, with phi = e^theta/(1+e^theta),
f(y; phi) propto phi^y * (1-phi)^(1-y)
}
\examples{

hist(rcobin(1000, 2, 3), freq = FALSE)
xgrid = seq(0, 1, length = 500)
lines(xgrid, dcobin(xgrid, 2, 3))

}
