% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sample.rowwise}
\alias{sample.rowwise}
\title{Vectorized sampling (by row by row) with probability stored in rows of matrix
source: https://stackoverflow.com/questions/20508658/sampling-repeatedly-with-different-probability}
\usage{
sample.rowwise(Weight)
}
\arguments{
\item{Weight}{n by K matrix, each row corresponds to unnormalized probabilities of sampling weights}
}
\value{
integers with length n, each element in {1,...,K}
}
\description{
Vectorized sampling (by row by row) with probability stored in rows of matrix
source: https://stackoverflow.com/questions/20508658/sampling-repeatedly-with-different-probability
}
