% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmicobinreg.R
\name{spmicobinreg}
\alias{spmicobinreg}
\title{Title}
\usage{
spmicobinreg(
  formula,
  data,
  link = "cobit",
  coords,
  lonlat = FALSE,
  NNGP = FALSE,
  contrasts = NULL,
  priors = list(beta_intercept_scale = 10, beta_scale = 2.5, beta_df = Inf),
  nngp.control = list(n.neighbors = 15, ord = order(coords[, 1])),
  nburn = 1000,
  nsave = 1000,
  nthin = 1
)
}
\arguments{
\item{nthin}{}
}
\description{
Title
}
