% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobinreg.R
\name{cobinreg}
\alias{cobinreg}
\title{cobin generalized linear (mixed) models}
\usage{
cobinreg(
  formula,
  data,
  link = "cobit",
  contrasts = NULL,
  priors = list(beta_intercept_scale = 10, beta_scale = 2.5, beta_df = Inf),
  nburn = 1000,
  nsave = 1000,
  nthin = 1,
  MH = F,
  lambda_fixed = NULL
)
}
\description{
cobin generalized linear (mixed) models
}
