% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/micobinreg.R
\name{micobinreg}
\alias{micobinreg}
\title{micobin generalized linear (mixed) models}
\usage{
micobinreg(
  formula,
  data,
  link = "cobit",
  contrasts = NULL,
  priors = list(beta_intercept_scale = 10, beta_scale = 2.5, beta_df = Inf),
  nburn = 1000,
  nsave = 1000,
  nthin = 1
)
}
\description{
micobin generalized linear (mixed) models
}
